API: Remove .mend_slices()
There were some concerns that these interfaces were not sound.

This functionality was only intended to "mend" just-recently-split
slices, but there'd need to be a guarantee that the only use was like
that, for it to be safe.

I didn't immediately find a way to wrap it up into a cool safe package.